aoc-solutions/src/test/java/org/thermoweb/aoc/days/Day14Test.java
package org.thermoweb.aoc.days;

import static org.junit.jupiter.api.Assertions.assertEquals;

import java.lang.Exception;
import java.math.BigInteger;
import java.util.Optional;
import org.junit.jupiter.api.Test;
import org.thermoweb.aoc.Day;
import org.thermoweb.aoc.DayRunner;

class Day14Test {
    private final Day day = new Day14();

    @Test
    void test_part_one() throws Exception {
        assertEquals(Optional.of(BigInteger.valueOf(8108)), day.partOne(DayRunner.getExample(14)));
    }

    @Test
    void test_part_two() throws Exception {
        assertEquals(Optional.of(BigInteger.valueOf(1242)), day.partTwo(DayRunner.getExample(14)));
    }
}
